+++ NEWS.md
@@ -11,6 +11,7 @@ from the US EPA, AirNow, AIRSIS, WRCC and others.
 
 ### PWFSLSmoke 1.0.19
 
+ * fixed single-digit month parsing in `airnow_load()`
  * capitalized `AQI$names`: 'Good', 'Moderate', ...
  * added `addAQIStackedBars()` function
  * added example test for `monitor_load()` which then tests `monitor_combine()`